# services/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "0.0.1",
        "title": "Users Service",
        "description": "Swagger spec for documenting the users service"
    },
    "servers": [
        {
            "url": "http://localhost"
        }
    ],
    "paths": {
        "/users/ping": {
            "get": {
                "summary": "Just a sanity check",
                "responses": {
                    "200": {
                        "description": "Will return 'pong!'"
                    }
                }
            }
        },
        "/users": {
            "get": {
                "summary": "Returns all users",
                "responses": {
                    "200": {
                            "description": "user object"
                    }
                }
            },
            "post": {
                "summary": "Adds a new user",
                "requestBody": {
                    "description": "User to add",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/user-full"
                            }
                        }
                    }
                },
                "security": [
                    {
                    "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User added"
                    }
                }
            }

        },
        "/users/{id}": {
            "get": {
                "summary": "Returns a user based on a single user ID",
                "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "description": "ID of user to fetch",
                    "required": true,
                    "schema": {
                    "type": "integer",
                    "format": "int64"
                    }
                }
                ],
                "responses": {
                    "200": {
                    "description": "user object"
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
            "summary": "Creates a new user",
            "requestBody": {
                "description": "User to add",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/user-full",
                            "type": "object",
                            "required": [
                            "username",
                            "email",
                            "password"
                        ],
                        "properties": {
                            "username": {
                            "type": "string"
                        },
                        "email": {
                            "type": "string"
                        },
                        "password": {
                            "type": "string"
                        }
                        }
                        }
                    }
                }
            },
            "responses": {
                "200": {
                    "description": "user object"
                }
            }
            }
        },
        "/auth/login": {
            "post": {
                "summary": "Logs a user in",
                "requestBody": {
                    "description": "User to log in",
                    "required": true,
                    "content": {
                        "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/user"
                        }
                        }
                    }
                },
            "responses": {
                "200": {
                    "description": "Successfully logged in"
                }
            }
            }
        },
        "/auth/status": {
            "get": {
                "summary": "Returns the logged in user's status",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
            "responses": {
                "200": {
                    "description": "user object"
                }
            }
            }
        },
        "/auth/logout": {
            "get": {
                "summary": "Logs a user out",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
            "responses": {
                "200": {
                    "description": "Successfully logged out"
                }
            }
            }
        }


    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
            "type": "http",
            "scheme": "bearer"
        },
        "schemas": {
            "user": {
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
            "user-full": {
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            }
        }
        }
    }
}